feat: adds new Homepage
